+++ src/Exception/SettingException.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace SettingsManager\Exception;
+
+interface SettingException
+{
+    /* pass */
+}