+++ tests/Provider/DefaultValuesProviderTest.php
@@ -1,0 +1,29 @@
+<?php
+
+namespace SettingsManager\Provider;
+
+use SettingsManager\AbstractSettingsProviderTest;
+use SettingsManager\Contract\SettingProviderInterface;
+
+class DefaultValuesProviderTest extends AbstractSettingsProviderTest
+{
+    /**
+     * @return array  Keys are setting names, values are setting values
+     */
+    protected function getExpectedValues(): array
+    {
+        return [
+            'string_setting' => 'test',
+            'integer_setting' => 3,
+            'decimal_setting' => 25.3
+        ];
+    }
+
+    /**
+     * @return SettingProviderInterface
+     */
+    protected function getProviderInstance(): SettingProviderInterface
+    {
+        return new DefaultValuesProvider($this->buildDefinitions());
+    }
+}